refactor(#42 WI-10b): split ReadiumEPUBHost body/lifecycle into +Body (300-line budget)
The host accreted WIs 5-12 + the WI-7/8/10b refinements to 323 lines. Move
coreBody / readyNavigator / openHostTask / onHostDisappear into a new
ReadiumEPUBHost+Body.swift extension (host 323→145, +Body 194 — both <300).
viewModel/restoredLocator @State drop `private` so the extension can read them.
No behavior change.
